+++ webtack/jsinterop-generator/src/test/fixtures/global_interface/output/gwt/test/java/com/example/TransferableTestCompile.java
@@ -1,0 +1,8 @@
+package com.example;
+
+import javax.annotation.Generated;
+
+@Generated("org.realityforge.webtack")
+public final class TransferableTestCompile {
+  static Transferable $typeReference$;
+}