+++ webtack/jsinterop-generator/src/test/fixtures/global_interface/output/j2cl/test/java/com/example/WindowGlobalTestCompile.java
@@ -1,5 +1,6 @@
 package com.example;
 
+import javaemul.internal.annotations.DoNotAutobox;
 import javax.annotation.Generated;
 
 @Generated("org.realityforge.webtack")
@@ -38,6 +39,21 @@ public final class WindowGlobalTestCompile {
 
   public static Object get(final String name) {
     return WindowGlobal.get( name );
+  }
+
+  public static void postMessage(@DoNotAutobox final java.lang.Object message,
+      final String targetOrigin, final JsArray<Transferable> transfer) {
+    WindowGlobal.postMessage( message, targetOrigin, transfer );
+  }
+
+  public static void postMessage(@DoNotAutobox final java.lang.Object message,
+      final String targetOrigin, final Transferable[] transfer) {
+    WindowGlobal.postMessage( message, targetOrigin, transfer );
+  }
+
+  public static void postMessage(@DoNotAutobox final java.lang.Object message,
+      final String targetOrigin) {
+    WindowGlobal.postMessage( message, targetOrigin );
   }
 
   public static String id() {