+++ webtack/jsinterop-generator/src/test/fixtures/namespaces/output/j2cl/main/java/com/example/Console.java
@@ -2,8 +2,10 @@ package com.example;
 
 import javaemul.internal.annotations.DoNotAutobox;
 import javax.annotation.Generated;
+import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import jsinterop.annotations.JsMethod;
+import jsinterop.annotations.JsOverlay;
 import jsinterop.annotations.JsPackage;
 import jsinterop.annotations.JsType;
 
@@ -46,4 +48,35 @@ public final class Console {
    * @see <a href="https://developer.mozilla.org/en-US/docs/Web/API/console/warn">console.warn - MDN</a>
    */
   public static native void warn(@DoNotAutobox @Nullable Object... data);
+
+  /**
+   * Last argument is optional and a varargs candidate
+   */
+  public static native void table(@DoNotAutobox @Nullable Object tabularData,
+      @Nonnull JsArray<String> properties);
+
+  /**
+   * Last argument is optional and a varargs candidate
+   */
+  @JsOverlay
+  public static final void table(@DoNotAutobox @Nullable final Object tabularData,
+      @Nonnull final String... properties) {
+    _table( tabularData, properties );
+  }
+
+  @JsMethod(
+      name = "table"
+  )
+  private static native void _table(@DoNotAutobox @Nullable Object tabularData,
+      @Nonnull String[] properties);
+
+  /**
+   * Last argument is optional and a varargs candidate
+   */
+  public static native void table(@DoNotAutobox @Nullable Object tabularData);
+
+  /**
+   * Last argument is optional and a varargs candidate
+   */
+  public static native void table();
 }