+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ### Unreleased
 
+* Convert operations on global types and namespaces into varargs method in the jsinterop binding if the last argument is an array type. This modifies the following methods:
+  * `Console.table(...)`
+  * `DedicatedWorkerGlobal.postMessage(...)`
+  * `WindowGlobal.postMessage(...)`
+
 ### [v0.26](https://github.com/akasha/akasha/tree/v0.26) (2021-09-28) · [Full Changelog](https://github.com/spritz/spritz/compare/v0.25...v0.26) · [Source Diff](https://github.com/akasha/akasha-java/compare/v0.25...v0.26)
 
 Changes in this release: